feat(tsyringe): dependency injection
Started to configure tsyringe. create-user-service will use dependency injection from now on.
Coded an 'skeleton' for users-repository, but its not fully implemented
yet.

I had to import reflect-metadata on the test so it could run. tsyringe
needs it, so I've imported it at server.ts, but I had to import it on
the test for some reason.