fix(jwt-provider): secret cannot be undefined
Secret is imported from .env, so typescript recognizes it as string | undefined.
That throws an error because secret must be a string.
I fixed it with an ! to guarantee it will be a string, but eslint has a rule disallowing it.
So, instead of adding an !, I've just put the code inside an if block that validates the secret.